Refactor ChannelHintsUpdater
